backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.1.0
httpx==0.28.1
supabase==2.13.0
bcrypt==4.3.0
pyjwt==2.12.1
